+++ java/src/main/resources/META-INF/services/io.cucumber.messages.ndjson.JsonProvider
