+++ java/src/test/java/io/cucumber/messages/ndjson/test/AcceptanceTest.java
@@ -3,6 +3,7 @@ package io.cucumber.messages.ndjson.test;
 import io.cucumber.messages.MessageToNdjsonWriter;
 import io.cucumber.messages.NdjsonToMessageReader;
 import io.cucumber.messages.ndjson.Deserializer;
+import io.cucumber.messages.ndjson.Json;
 import io.cucumber.messages.ndjson.Serializer;
 import io.cucumber.messages.types.Envelope;
 import org.junit.jupiter.api.io.TempDir;
@@ -37,7 +38,7 @@ final class AcceptanceTest {
     }
 
     static Stream<Arguments> services() {
-        return JsonUtil.instances()
+        return Json.instances()
                 .map(json -> Arguments.argumentSet(
                         json.getClass().getSimpleName(),
                         json.deserializer(Envelope.class),
